glsl: add packing rules for tessellation stages to the packing code
Following patches will allow packing of varyings with explicit locations
via the component layout qualifier. Adding the rules here will enable
us to call an alternate path for packing tessellation stages with
explicit locations.

V3: Don't remove rules from the varying linking code as we need to
disable packing in the producer if the consumer is a TCS.

V2: move the tessellation packing rules, allow TES output to be packed.

Reviewed-by: Anuj Phogat <anuj.phogat@gmail.com>