glsl: lower tessellation varyings packed with component layout qualifier
For tessellation shaders we cannot just copy everything to the packed
varyings like we do in other stages as tessellation uses shared memory for
varyings, therefore it is only safe to copy array elements that the shader
actually uses.

This class searches the IR for uses of varyings and then creates
instructions that copy those vars to a packed varying. This means it is
easy to end up with duplicate copies if the varying is used more than once,
also arrays of arrays create a duplicate copy for each dimension that
exists. These issues are not easily resolved without breaking various
corner cases so we leave it to a later IR stage to clean up the mess.

Note that neither GLSL IR nor NIR can currently clean up the
duplicates when and indirect is used as an array index. This patch
assumes that NIR will eventually be able to clean this up.

V4: Fix IR validation for vector components accessed by an array style
subscript. Also add spec quote and link to spec bug to support not
writing vecotr component individually.

V3: Use correct base location for per-patch outputs, fix some comments,
don't segfault when using subscript to access vector components, and
for now don't support packing arrays of different size or varyings
that start at different locations in tcs out/tes in varyings.

V2: clone array index when creating array dereference for the packed
varying